+++ projects/core/src/entity.ts
@@ -11,7 +11,7 @@ import type {
   FindOptions,
   LifecycleEvent,
 } from './remult3/remult3.js'
-import type { FilterInfo } from './filter/filter-interfaces.js'
+import type { FilterPreciseValues } from './filter/filter-interfaces.js'
 
 export interface EntityOptions<entityType = any> {
   /**A human readable name for the entity */
@@ -72,10 +72,10 @@ export interface EntityOptions<entityType = any> {
    * @example
    * ```typescript
    * @Entity<Task>("tasks", {
-   *   apiPreprocessFilter: async (filter, { getFilterInfo }) => {
+   *   apiPreprocessFilter: async (filter, { getPreciseValues }) => {
    *     // Ensure that users can only query tasks for specific customers
-   *     const info = await getFilterInfo();
-   *     if (!info.preciseValues.customerId) {
+   *     const preciseValues = await getPreciseValues();
+   *     if (!preciseValues.customerId) {
    *       throw new ForbiddenError("You must specify a valid customerId filter");
    *     }
    *     return filter;
@@ -85,7 +85,7 @@ export interface EntityOptions<entityType = any> {
    */
   apiPreprocessFilter?: (
     filter: EntityFilter<entityType>,
-    info: PreprocessFilterInfo<entityType>,
+    info: PreprocessFilterEvent<entityType>,
   ) => EntityFilter<entityType> | Promise<EntityFilter<entityType>>
 
   /**
@@ -98,7 +98,7 @@ export interface EntityOptions<entityType = any> {
    */
   backendPreprocessFilter?: (
     filter: EntityFilter<entityType>,
-    info: PreprocessFilterInfo<entityType>,
+    info: PreprocessFilterEvent<entityType>,
   ) => EntityFilter<entityType> | Promise<EntityFilter<entityType>>
 
   /** A filter that will be used for all queries from this entity both from the API and from within the backend.
@@ -215,20 +215,20 @@ export interface EntityOptions<entityType = any> {
  * Provides additional information and utilities for preprocessing filters in API and backend operations.
  * @template entityType The type of the entity being filtered.
  */
-export interface PreprocessFilterInfo<entityType> {
+export interface PreprocessFilterEvent<entityType> {
   /**
    * Metadata of the entity being filtered.
    */
   metadata: EntityMetadata<entityType>
 
   /**
-   * Retrieves filter information for a given filter or the current filter being preprocessed if no filter is provided.
+   * Retrieves precise values for each property in a filter for an entity.
    * @param filter Optional filter to analyze. If not provided, the current filter being preprocessed is used.
-   * @returns A promise that resolves to a FilterInfo object containing the filter information.
+   * @returns A promise that resolves to a FilterPreciseValues object containing the precise values for each property.
    
-  * {@Link FilterInfo }
+  * {@Link FilterPreciseValues }
    */
-  getFilterInfo(
+  getFilterPreciseValues(
     filter?: EntityFilter<entityType>,
-  ): Promise<FilterInfo<entityType>>
+  ): Promise<FilterPreciseValues<entityType>>
 }