+++ projects/core/src/remult3/remult3.ts
@@ -814,8 +814,7 @@ export type ClassFieldDecorator<entityType, valueType> = (
 ) => void
 
 //p1 - add update many and delete many to the graphql
-//p1 - remove filter info and return presciseValues
-//p1 - PreprocessFilterInfo => PreprocessFilterEvent
+
 //y1 - JY suggestion of having generate as part of the ensure schema - not sure
 //y1 - currently for string fields we default for 255 in knex (which is their default) why not just use text (varchar max) - and only use exact length when one is provided?
 //y1 - consider changing the sql raw filter to return the sql, instead of setting the sql property